backend/requirements.txt
Django
djangorestframework
drf-nested-routers
django-cors-headers
djangorestframework-simplejwt
pika
Pillow